+++ pytsite/console/__init__.py
@@ -5,8 +5,8 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 # Public API
-from ._api import register_command, get_command, run_command, usage, run, print_normal, print_info, print_error, \
-    print_success, print_warning
+from ._api import register_command, get_command, get_current_command, run_command, usage, run, print_normal, \
+    print_info, print_error, print_success, print_warning
 
 from pytsite import lang as _lang
 from ._command import Command